+++ test/run_pass/test_process_exit_child_exit.js
