+++ test/run_pass/test_process_exit_handler.js
@@ -1,5 +1,15 @@
 'use strict';
+var assert = require('assert');
+var common = require('../common');
 
-process.on('exit', function() {
-  process.exit(0);
+var fn = (c) => {
+  assert.strictEqual(0, c);
+};
+
+process.on('exit', function(c) {
+  console.log(c)
+  process.exitCode = 123
+  fn(c);
 });
+
+fn = common.mustCall(fn);
