+++ test/run_pass/test_process_exit_immediately.js
@@ -1,6 +1,6 @@
 'use strict';
 var assert = require('assert');
 
-process.exit(233);
+process.exit(0);
 
 assert.fail();
