+++ test/run_pass/test_process_exit_ticks.js
@@ -17,4 +17,4 @@ process.nextTick(function() {
   assert.fail();
 });
 
-process.exit(233);
+process.exit(0);
